pyproject.toml
[project]
name = "arclio-login"
version = "1.0.0"
description = "Arclio CLI for Kinde OAuth authentication"
readme = "README.md"
license = { text = "Proprietary" }
authors = [{ name = "Arclio" }]
requires-python = ">=3.11"
keywords = ["arclio", "kinde", "oauth", "authentication", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
arclio = "arclio_login.cli:main"

[project.urls]
Homepage = "https://github.com/Arclio/arclio-login"
Repository = "https://github.com/Arclio/arclio-login"

[dependency-groups]
dev = [
    "poethepoet>=0.34.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.408",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arclio_login"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
include = ["src"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv/**",
    ".pytest_cache/**",
]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
reportMissingImports = "error"
reportGeneralTypeIssues = "warning"

[tool.poe.tasks]
pyright = "uv run pyright ."
lint = "uv run ruff check . --fix"
format = "uv run ruff format ."
pre-commit = ["format", "lint", "pyright"]
test = "uv run pytest tests/ -v"
test-cov = "uv run pytest tests/ -v --cov=src/arclio_login --cov-report=term-missing"

[tool.poe.tasks.setup-hooks]
shell = "uv run pre-commit install"
help = "Install pre-commit hooks (required once after cloning)"
